# Fix perf regression in traversal