Due to the conversion from a recursive traversal to an iterator, the load_node function was called many times, causing perf problems. This fixes this by caching that call and doing a few other things to optimize traversal.

There is a test that makes sure we only request a small number of blocks when drilling down to a single value.

Traversal with a query that matches everything is now a tiny bit slower than collect, and traversal of a tiny section of the tree is fast as expected.

Still something weird that deserves looking into:

```
running 1 test
collect 65 reads, 236798us
iter    65 reads, 282631us
small   35 reads, 131453us
tiny    35 reads, 110551us < this should not be 35 reads...
test ops_count_1 ... ok
```

Now this is looking better:

```
create 1.372394
collect 0.2854
filter_common 0.176785
filter_rare 0.084883
```